% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_paths.R
\name{all_res_file_paths}
\alias{all_res_file_paths}
\title{List paths for all result files}
\usage{
all_res_file_paths(study, repls, opmods, estmods, root_dir = ".")
}
\arguments{
\item{study}{Study name}

\item{repls}{Replicate numbers}

\item{opmods}{Operating model numbers}

\item{estmods}{Estimation model names}

\item{root_dir}{Root directory}
}
\value{
List where each element has \code{rdata} and \code{indexcsv} paths
}
\description{
List all paths to results files for a study.
}
\author{
John K Best
}
