% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_spatq.R
\name{attach_optdiags}
\alias{attach_optdiags}
\title{Attach optimization diagnostics}
\usage{
attach_optdiags(newfit, oldfit = NULL, obj)
}
\arguments{
\item{newfit}{new optimization result}

\item{oldfit}{previous optimization result}

\item{obj}{TMB objective function}
}
\value{
A \code{spatq_fit} object
}
\description{
Attach optimization diagnostics comparing a previous optimization result to
a new optimization result.
}
\author{
John K Best
}
