% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_expr_str}
\alias{build_expr_str}
\title{Build an expression string from a list of string vectors}
\usage{
build_expr_str(parts_list)
}
\arguments{
\item{parts_list}{List of string vectors}
}
\value{
Single string pasted together
}
\description{
Paste first elements together in order, then second elements in reverse
order.
}
\author{
John K Best
}
