% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{check_scenario}
\alias{check_scenario}
\title{Check scenario name}
\usage{
check_scenario(sc)
}
\arguments{
\item{sc}{Scenario name, as a string}
}
\value{
\code{TRUE} if scenario is correct
}
\description{
Check that scenario name is correct. Available options are "pref", "spat",
or "combo". Any other string will throw an error.
}
\author{
John Best
}
