% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{create_index_df}
\alias{create_index_df}
\title{Create a data frame with index locations and times}
\usage{
create_index_df(step = 1, T = 25)
}
\arguments{
\item{step}{Step in the spatial domain}

\item{T}{Number of years}
}
\value{
A \code{tibble} with columns \code{s1}, \code{s2}, and \code{time}
and attribute \code{step} for calculating the integration weights.
}
\description{
Create a data frame with a regular grid over the spatial domain for each
year of the index.
}
\author{
John Best
}
