% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_spatq.R
\name{dim_or_len}
\alias{dim_or_len}
\title{Get \code{dim} if available, \code{length} otherwise}
\usage{
dim_or_len(x)
}
\arguments{
\item{x}{Object to be \code{dim}d}
}
\value{
Vector (possible length 1) of the dimensions of \code{x}.
}
\description{
Get \code{dim} if available, \code{length} otherwise
}
