% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{domain_boundary}
\alias{domain_boundary}
\title{Define the domain boundary}
\usage{
domain_boundary()
}
\value{
An \code{inla.mesh.segment} giving the boundary of the 100 by 100
domain
}
\description{
Assumes the simulation \eqn{[0, 100]×[0, 100]} domain.
}
\author{
John Best
}
