% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_specs.R
\name{em_subsample}
\alias{em_subsample}
\title{Specify data subsample for estimation model}
\usage{
em_subsample(estmod, n_comm = 4000)
}
\arguments{
\item{estmod}{Estimation model name}

\item{n_comm}{Number of commercial observations to use}
}
\value{
A data frame to be used by \code{\link{subsample_catch}}
}
\description{
Return a data frame compatible with \code{\link{subsample_catch}} specifying
the number of catch observations to use for each estimation model.
}
\details{
Valid estimation model names that subset observations to use only survey data are:

\itemize{
\item \code{"design"}
\item \code{"model"}
\item \code{"survey"}
\item \code{"survey_spt"}
}

Estimation models that include all survey observations and \code{n_comm} commercial observations
are:
\itemize{
\item \code{"design_all"}
\item \code{"model_all"}
\item \code{"spatial_ab"}
\item \code{"spatial_q"}
\item \code{"sptemp_ab"}
\item \code{"sptemp_q"}
}
}
\author{
John K Best
}
