% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_plots.R
\name{field_image}
\alias{field_image}
\title{Plot a field}
\usage{
field_image(z, col = hcl.colors(12), ...)
}
\arguments{
\item{z}{Field matrix}

\item{col}{Colormap}

\item{...}{Additional arguments to pass to \code{\link[graphics]{image}}}
}
\description{
Plot the \code{\link[graphics]{image}} of a random field, without axes and
with aspect ratio one.
}
\author{
John K Best
}
