% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_spatq.R
\name{fit_spatq}
\alias{fit_spatq}
\alias{report_spatq}
\alias{sdreport_spatq}
\title{Fit a spatq model by maximum likelihood}
\usage{
fit_spatq(obj, ..., method = "BFGS")

report_spatq(obj)

sdreport_spatq(obj)
}
\arguments{
\item{obj}{A spatq \code{ADFun}, as returned by \code{prepare_adfun} or
\code{make_sim_adfun}}

\item{...}{Additional options to pass to \code{optim}, e.g. \code{control}}

\item{method}{Optimization method to uses (see \code{optim} for available
options)}
}
\value{
An optimization object, report list, or sdreport list
}
\description{
Use \code{optim} to estimate parameters of a \code{spatq} model.
}
\section{Functions}{
\itemize{
\item \code{report_spatq}: Get object report

\item \code{sdreport_spatq}: Get object sdreport
}}

\author{
John Best
}
