% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_spatq.R
\name{fix_nlminb_fit}
\alias{fix_nlminb_fit}
\title{Make \code{nlminb} output match \code{optim} output}
\usage{
fix_nlminb_fit(fit)
}
\arguments{
\item{fit}{\code{\link[stats]{nlminb}} output}
}
\value{
Optimization result that matches \code{\link[stats]{optim}}
}
\description{
Reorders output, renames "objective" to "value", and drops "iterations".
}
\author{
John K Best
}
