% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{fixpar_corr}
\alias{fixpar_corr}
\title{Correlation matrix of the fixed-effect parameters}
\usage{
fixpar_corr(sdr)
}
\arguments{
\item{sdr}{An \code{\link[TMB]{sdreport}} from a TMB model fit}
}
\value{
The correlation matrix of the fixed-effect parameters.
}
\description{
Extract the covariance matrix from an \code{\link[TMB]{sdreport}} and
convert it to a correlation matrix.
}
\author{
John K Best
}
