% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{fixpar_covcond}
\alias{fixpar_covcond}
\title{Calculate the condition number of the fixed-effect parameter
covariance matrix}
\usage{
fixpar_covcond(sdr, ...)
}
\arguments{
\item{sdr}{An \code{\link[TMB]{sdreport}} from a TMB model fit}

\item{...}{Additional arguments to pass to \code{\link[base]{kappa}}}
}
\value{
The condition number of fixed-effect parameter covariance matrix.
}
\description{
Larger values indicate that this matrix is "more singular", and so more
susceptible to numerical error when inverted.
}
\author{
John K Best
}
