% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{fixpar_grad}
\alias{fixpar_grad}
\title{Gradients at the fitted values}
\usage{
fixpar_grad(sdr)
}
\arguments{
\item{sdr}{An \code{\link[TMB]{sdreport}} from a TMB model fit}
}
\value{
Named vector of gradients
}
\description{
Get the gradients at the estimated values of the fixed effects.
}
\author{
John K Best
}
