% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{fixpar_highcorr}
\alias{fixpar_highcorr}
\title{Fixed-effect parameters with high correlations}
\usage{
fixpar_highcorr(sdr, corr_cutoff = 0.8)
}
\arguments{
\item{sdr}{An \code{\link[TMB]{sdreport}} from a TMB model fit}

\item{corr_cutoff}{Return pairs of parameters with correlations larger than
this cutoff}
}
\value{
A data frame with columns \code{par1}, \code{par2}, and \code{cor}
}
\description{
Return a data frame with parameter pairs that have a higher absolute
correlation than a given cutoff.
}
\details{
Note that the returned data frame will not include duplicates. Parameters
with repeated names are disambiguated using \code{\link[base]{make.unique}}, so
they are essentially zero-indexed.
}
\author{
John K Best
}
