% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{fixpar_mgc}
\alias{fixpar_mgc}
\title{Maximum gradient component at fitted parameter values}
\usage{
fixpar_mgc(sdr, abs = TRUE)
}
\arguments{
\item{sdr}{An \code{\link[TMB]{sdreport}} from a TMB model fit}

\item{abs}{Return the maximum absolute gradient or maximum gradient?}
}
\value{
Largest gradient with name of parameter
}
\description{
Find the largest gradient at the fitted values.
}
\author{
John K Best
}
