% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{fixpar_nulleigs}
\alias{fixpar_nulleigs}
\title{Find the parameters that contribute to a singular covariance matrix}
\usage{
fixpar_nulleigs(sdr, reltol = 1e-04, abstol = NULL)
}
\arguments{
\item{sdr}{An \code{\link[TMB]{sdreport}} object from a TMB model}

\item{reltol}{Relative tolerance for small eigenvalues}

\item{abstol}{Absolute tolerance for small eigenvalues; takes precedence}
}
\value{
A list with \code{values} (eigenvalues), \code{vectors}
(eigenvectors), and \code{sortedpars}, which is a matrix of parameter
names with each column ordered by the corresponding loadings of the
eigenvectors (largest loadings at the top)
}
\description{
Take the eigendecomposition of the fixed-effect covariance matrix, filter
out the smallest eigenvalues, and then order the parameter names by the
loadings in the corresponding eigenvectors.
}
\details{
The \code{reltol} argument sets a minimum for the ratio of each eigenvalue
to the largest eigenvalue. The \code{abstol} argument sets an absolute
minimum value. The latter takes precedence if it is passed.
}
\author{
John K Best
}
