% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{fixpar_sd}
\alias{fixpar_sd}
\title{Return the standard deviations of the fixed-effect parameters}
\usage{
fixpar_sd(sdr)
}
\arguments{
\item{sdr}{An \code{\link[TMB]{sdreport}} from a TMB model fit}
}
\value{
The named vector of standard deviations
}
\description{
Return the square root of the diagonal elements of the fixed-effect
parameter covariance matrix.
}
\author{
John K Best
}
