% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gather_nvec}
\alias{gather_nvec}
\title{Gather a named vector into a list}
\usage{
gather_nvec(vec)
}
\arguments{
\item{vec}{Named vector}
}
\value{
List with elements being a unique name in the input vector
}
\description{
Given a named vector, produce a list that gathers the elements of the vector
with the same name into a list where each element takes these unique names.
}
\examples{
vec <- c(1, 1, 1, 1, 2, 2, 2, 3, 3, 4)
names(vec) <- c("a", "a", "a", "a", "b", "b", "b", "c", "c", "d")
gather_nvec(vec)

}
\author{
John Best
}
