% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_spde.R
\name{generate_aniso_fem}
\alias{generate_aniso_fem}
\title{Generate anisotropic FEM matrices}
\usage{
generate_aniso_fem(mesh)
}
\arguments{
\item{mesh}{INLA mesh}
}
\value{
List of data expected for \code{spde_aniso_t} TMB object
}
\description{
Generate FEM matrices for anisotropic SPDE. Code adapted from TMB anisotropy
example.
}
\author{
John Best
}
