% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{generate_empty_projection}
\alias{generate_empty_projection}
\title{Generate an empty project matrix}
\usage{
generate_empty_projection(mesh, data_df, group = NULL)
}
\arguments{
\item{mesh}{INLA mesh to project}

\item{data_df}{Data frame with \code{s1} and \code{s2} columns as
coordinates to project to}

\item{group}{Year of observation for spatiotemporal effect}

\item{...}{Not used but availble for consistent calls with
\code{generate_projection}}
}
\value{
Sparse \code{Matrix::Matrx} of the appropriate dimensions, but
filled with zeros
}
\description{
When spatial or spatiotemporal effects are map'd to zero, pass an empty
projection matrix. Not exported, use \code{zero = TRUE} in \code{generate_projection}
instead.
}
\author{
John Best
}
