% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{generate_fem}
\alias{generate_fem}
\title{Generate FEM matrices}
\usage{
generate_fem(mesh)
}
\arguments{
\item{mesh}{INLA mesh to generate FEM matrices from}
}
\value{
List of sparse matrices suitable for passing to TMB.
}
\description{
Generate the FEM matrices with appropriate names to pass to TMB as a
\code{spde_t}.
}
\author{
John Best
}
