% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{get_coordref}
\alias{get_coordref}
\title{Read the index to coordinate reference table}
\usage{
get_coordref(root_dir = ".")
}
\arguments{
\item{root_dir}{Directory containing the file "coordref.csv"}
}
\value{
A \code{tibble} with columns \code{loc_idx}, \code{s1}, and
\code{s2}.
}
\description{
Get the table of location indices and corresponding location coordinates
}
\details{
Assumes that the table is in the file \code{coordref.csv} in the working
directory, as saved by \code{FisherySim.jl}
}
\author{
John Best
}
