% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_fits.R
\name{get_index_filetype}
\alias{get_index_filetype}
\title{Get index file type}
\usage{
get_index_filetype(spec)
}
\arguments{
\item{spec}{A \code{list} or \code{spatq_simstudyspec}}
}
\value{
Either string \code{"feather"} or \code{"csv"}
}
\description{
Checks whether the Feather or CSV index result files exist, and returns the
appropriate string. If both exist, returns the file type that was most
recently modified.
}
\author{
John K Best
}
