% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{get_newpars}
\alias{get_newpars}
\title{Get updated parameter values}
\usage{
get_newpars(obj)
}
\arguments{
\item{obj}{object to extract new parameter values from}
}
\value{
named list with parameter values to update
}
\description{
Extract updated parameter values from a previous fit, provided as a named
vector, list, \code{spatq_fit}, or \code{spatq_obj} that has already been
optimized.
}
\author{
John K Best
}
