% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_paths.R
\name{index_path}
\alias{index_path}
\title{Construct the relative path to the index file output by a fit.}
\usage{
index_path(spec, filetype)
}
\arguments{
\item{spec}{A \code{\link{spatq_simstudyspec}} object}

\item{filetype}{Index file type, either "feather"  or "csv"}
}
\value{
The relative path to the index data file
}
\description{
Construct the relative path to the index file output by a fit.
}
\author{
John K Best
}
