% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{init_fixef}
\alias{init_fixef}
\title{Initial estimates of fixed effect parameters}
\usage{
init_fixef(setup, spec, ...)
}
\arguments{
\item{setup}{A \code{\link{spatq_setup}}}

\item{spec}{A model specification, as from \code{\link{specify_estimated}}}

\item{...}{Further arguments to pass to \code{\link{spatq_fit}}}
}
\value{
A \code{\link{spatq_setup}} with parameters of fixed-effects model
initialized to their MLEs
}
\description{
Fit the specified model without any random effects and use the resulting
parameter estimates as initial values.
}
\author{
John Best
}
