% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_plots.R
\name{interleave_estgen}
\alias{interleave_estgen}
\title{Alternate estimated and generative parameter vectors}
\usage{
interleave_estgen(est, gen = NULL, checklen = TRUE)
}
\arguments{
\item{est}{Named list of estimated parameters}

\item{gen}{named list of generative parameter vectors}

\item{checklen}{Check corresponding lengths are equal?}
}
\value{
A list with alternating elements of \code{est} and \code{gen}
}
\description{
Extracts the elements of \code{gen} with the names in \code{est}, then
alternates them so that the first element of the returned list is the first
from \code{est}, the second is the element of \code{gen} with the same name,
etc. Also adds the \code{"isgen"} attribute (logical) indicating whether
these are generative or estimated values.
}
\details{
Not symmetric; only elements with names in \code{est} will be kept. All
elements in \code{est} should be included in \code{gen}.
}
\author{
John K Best
}
