% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_plots.R
\name{isgen}
\alias{isgen}
\alias{isest}
\alias{isgen<-}
\alias{isest<-}
\title{Is this parameter generative or estimated?}
\usage{
isgen(x)

isest(x)

isgen(x) <- value

isest(x) <- value
}
\arguments{
\item{x}{Object}

\item{value}{Logical; is this a generative/estimated parameter?}
}
\value{
\code{x} with attribute \code{"isgen"} or logical
}
\description{
Set or get the status of a parameter vector; is it generative or estimated?
}
\section{Functions}{
\itemize{
\item \code{isest}: Is this object estimated?

\item \code{isgen<-}: Set whether this object is generative?

\item \code{isest<-}: Set attribute \code{"isgen"} to \code{!isest}
}}

\author{
John K Best
}
