% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_plots.R
\name{layout_mat}
\alias{layout_mat}
\alias{layout_1mat}
\alias{layout_indoffset}
\alias{layout_widths}
\title{Construct layout matrix for field plots}
\usage{
layout_mat(nyears, incl_gen = TRUE, npar = 1, colorbar = FALSE)

layout_1mat(nyears, incl_gen = TRUE, idxoffset = 0, colorbar = FALSE)

layout_indoffset(nyears, incl_gen = TRUE, npar = 1, colorbar = FALSE)

layout_widths(nyears, incl_gen = TRUE, npar = 1, colorbar = FALSE)
}
\arguments{
\item{nyears}{Number of years to plot}

\item{incl_gen}{Include generative fields?}

\item{npar}{Number of parameters to plot}

\item{colorbar}{Include colorbar?}

\item{idxoffset}{Largest of the previously used indices; this layout matrix
will start at \code{idxoffset + 1}.}
}
\value{
A layout matrix
}
\description{
Construct a matrix for \code{\link[graphics]{layout}} with parameters in
rows and years in columns.
}
\section{Functions}{
\itemize{
\item \code{layout_1mat}: Layout matrix for a single parameter

\item \code{layout_indoffset}: Vector of index offsets for each parameter in field
plot

\item \code{layout_widths}: Widths of field plots; thinner for colorbar if
present
}}

\author{
John K Best
}
