% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_spde.R
\name{loc_grid}
\alias{loc_grid}
\title{Generate a location grid}
\usage{
loc_grid(step = 1)
}
\arguments{
\item{step}{Distance between subsequent locations; defaults to 1}
}
\value{
A two-column matrix with of locations in two dimensions, each
\code{step} apart.
}
\description{
Generate the coordinates of a regular grid on a \eqn{[0, 100]×[0, 100]}
domain with steps in both direction of \code{step}.
}
\author{
John Best
}
