% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{make_sim_adfun}
\alias{make_sim_adfun}
\title{Create ADFun from simulated data}
\usage{
make_sim_adfun(repl, sc, sub_df = NULL, root_dir = ".", max_T = NULL)
}
\arguments{
\item{repl}{Replicate number}

\item{sc}{Scenario; "naive", "simple", "scaled", or "shared"}

\item{sub_df}{Data frame indicating subsampling strategy; see
\code{subsample_catch}}

\item{root_dir}{Directory to load data from}

\item{max_T}{Last year of data to include}
}
\value{
A TMB ADFun suitable for optimization
}
\description{
Read in a simulated data set and construct a TMB ADFun for model fitting.
From the working directory, replicates should be in subdirectories and
scenarios named appropriately. See \code{read_catch} for details.
}
\author{
John Best
}
