% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatq_result.R
\name{mgc}
\alias{mgc}
\title{Maximum gradient component}
\usage{
mgc(x, abs = TRUE)
}
\arguments{
\item{x}{\code{spatq_fit} or \code{spatq_result}}

\item{abs}{Use absolute value of gradient?}
}
\value{
Maximum gradient component of the fixed effect parameters
}
\description{
Maximum gradient component
}
\author{
John K Best
}
