% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{new_spatq_setup}
\alias{new_spatq_setup}
\alias{spatq_setup}
\title{Set up inputs for spatq model}
\usage{
new_spatq_setup(data, parameters, map, random)

spatq_setup(
  catch_df,
  spatq_spec,
  index_step,
  init_fixef = TRUE,
  mesh_resolution = "coarse",
  ...
)
}
\arguments{
\item{data}{Data for spatq model, as produced by \code{\link{prepare_data}}}

\item{parameters}{Parameter list as from \code{\link{prepare_pars}}}

\item{map}{Map list as from \code{\link{prepare_map}}}

\item{random}{Random vector as from \code{\link{prepare_random}}}

\item{catch_df}{Data frame with catch observations, as from
\code{\link{read_catch}}}

\item{spatq_spec}{Model specification from \code{\link{specify_estimated}}}

\item{index_step}{Index grid step size \code{\link{create_index_df}}}

\item{init_fixef}{Use parameter estimates from model with only fixed effects
as initial values?}

\item{mesh_resolution}{Mesh resolution to pass to
\code{\link{generate_mesh}}. Must be "coarse", "medium", or "fine".}

\item{...}{Additional arguments to pass to \code{\link{init_fixef}}, which
are then passed to \code{\link{spatq_fit}} for the fixed effect model fit}
}
\value{
A \code{spatq_setup} object ready to be used to construct a spatq
objective function
}
\description{
Checks that \code{data}, \code{parameters}, and \code{map} are consistent
then puts them in a list for use by \code{\link{prepare_adfun}}
}
\section{Functions}{
\itemize{
\item \code{spatq_setup}: Convenient constructor for model setups
}}

\author{
John K Best
}
