% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{new_spatqsetup}
\alias{new_spatqsetup}
\alias{spatqsetup}
\title{Set up inputs for spatq model}
\usage{
new_spatqsetup(data, parameters, map, random)

spatqsetup(catch_df, spatqspec, index_step)
}
\arguments{
\item{data}{Data for spatq model, as produced by \code{\link{prepare_data}}}

\item{parameters}{Parameter list as from \code{\link{prepare_pars}}}

\item{map}{Map list as from \code{\link{prepare_map}}}

\item{random}{Random vector as from \code{\link{prepare_random}}}

\item{catch_df}{Data frame with catch observations, as from
\code{\link{read_catch}}}

\item{spatqspec}{Model specification, \code{\link{specify_estimated}}}

\item{index_step}{Index grid step size \code{\link{create_index_df}}}
}
\value{
A \code{spatqsetup} object ready to be used to construct a spatq
objective function
}
\description{
Checks that \code{data}, \code{parameters}, and \code{map} are consistent
then puts them in a list for use by \code{\link{prepare_adfun}}
}
\section{Functions}{
\itemize{
\item \code{spatqsetup}: Convenient constructor for model setups
}}

\author{
John K Best
}
