% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{nulleigs}
\alias{nulleigs}
\alias{fixpar_nulleigs}
\alias{jointprec_nulleigs}
\title{Find the parameters that contribute to a singular covariance matrix}
\usage{
nulleigs(x, reltol = 1e-04, abstol = NULL, vectrans = abs)

fixpar_nulleigs(sdr, reltol = 1e-04, abstol = NULL)

jointprec_nulleigs(sdr, reltol = 1e-04, abstol = NULL)
}
\arguments{
\item{x}{A matrix}

\item{reltol}{Relative tolerance for small eigenvalues}

\item{abstol}{Absolute tolerance for small eigenvalues; takes precedence}

\item{vectrans}{Transformation function applied before sorting eigenvector
elements, usually either \code{abs} or \code{identity}}

\item{sdr}{An \code{\link[TMB]{sdreport}} object from a TMB model}
}
\value{
A list with \code{values} (eigenvalues), \code{vectors}
(eigenvectors), and \code{sortedpars}, which is a matrix of parameter
names with each column ordered by the corresponding loadings of the
eigenvectors (largest loadings at the top)
}
\description{
Take the eigendecomposition of the fixed-effect covariance matrix, filter
out the smallest eigenvalues, and then order the parameter names by the
loadings in the corresponding eigenvectors.
}
\details{
The \code{reltol} argument sets a minimum for the ratio of each eigenvalue
to the largest eigenvalue. The \code{abstol} argument sets an absolute
minimum value. The latter takes precedence if it is passed.
}
\section{Functions}{
\itemize{
\item \code{fixpar_nulleigs}: Find (nearly) null eigenvalues in the fixed parameter
covariance matrix

\item \code{jointprec_nulleigs}: Find (nearly) null eigenvalues in the joint precision
matrix
}}

\author{
John K Best
}
