% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{par_expr_parts}
\alias{par_expr_parts}
\title{Strings from parameter names to be expression-ified}
\usage{
par_expr_parts(piece, hat = FALSE)
}
\arguments{
\item{piece}{String, piece of parameter name}

\item{hat}{Hat for estimated parameter?}
}
\value{
Vector of length one or two with pieces of parameter expression
}
\description{
Strings from parameter names to be expression-ified
}
\author{
John K Best
}
