% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{pars_data}
\alias{pars_data}
\title{Parameter values from data elements}
\usage{
pars_data(data_el, T = 1)
}
\arguments{
\item{data_el}{Data element to take dimensions from}

\item{T}{Number of years}
}
\value{
A number, vector, or matrix of zeros of appropriate dimension and
type for that parameter of the \code{spatq} model
}
\description{
Generate an (all-zero) parameter starting value of appropriate dimension
from the corresponding data element
}
\author{
John Best
}
