% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{pars_kappa}
\alias{pars_kappa}
\title{Calculate kappa from rho}
\usage{
pars_kappa(rho)
}
\arguments{
\item{rho}{Correlation range parameter}
}
\value{
The value of kappa corresponding to rho
}
\description{
The SPDE construction uses parameters kappa and tau, while the usual Matern
parameters are sigma and rho. The kappa parameter is \code{sqrt(8) / rho}.
}
\author{
John Best
}
