% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_spde.R
\name{pars_kappa}
\alias{pars_kappa}
\alias{pars_tau}
\alias{pars_rho}
\alias{pars_sig}
\title{The SPDE construction uses parameters kappa and tau, while the usual Matern
parameters are sigma and rho. The kappa parameter is \code{sqrt(8) / rho}.}
\usage{
pars_kappa(rho)

pars_tau(sig, rho)

pars_rho(kappa)

pars_sig(tau, kappa)
}
\arguments{
\item{rho}{Correlation range parameter}

\item{sig}{Marginal standard deviation}

\item{kappa}{Correlation decay parameter}

\item{tau}{Ratio of precision and correlation range}
}
\value{
Parameter value
}
\description{
The SPDE construction uses parameters kappa and tau, while the usual Matern
parameters are sigma and rho. These functions allow converting between the
parameterizations.
}
\details{
The tau parameter rescales the precision matrix to a given marginal standard
deviation, while \code{kappa} controls the correlation decay rate. In particular,
\itemize{
\item \code{tau = 2 * sqrt(pi) * sig * kappa},
\item \code{kappa = sqrt(8) / rho},
\item \code{sig = tau / (2 * sqrt(pi) * kappa)}, and
\item \code{rho = sqrt(8) / kappa}.
}

All assume that the Matern smoothness parameter is 1 and the domain is
two-dimensional
}
\section{Functions}{
\itemize{
\item \code{pars_kappa}: Calculate kappa from rho

\item \code{pars_rho}: Calculate the correlation range

\item \code{pars_sig}: Calculate the marginal standard deviation
}}

\author{
John Best
}
