% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{parse_coords}
\alias{parse_coords}
\title{Parse a tuple of numbers read as a string}
\usage{
parse_coords(coord_tuple)
}
\arguments{
\item{coord_tuple}{A string of the form \code{"(1.0, 2.0)"}}
}
\value{
The two numbers in the tuple
}
\description{
Parse coordinates that were saved as a tuple of \code{Float64} and read in
as a string. Much slower than using \code{get_coordref} and doing a
\code{join}.
}
\author{
John Best
}
