% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{plot_colorbar}
\alias{plot_colorbar}
\alias{plot_corcolorbar}
\title{Plot a colorbar legend}
\usage{
plot_colorbar(zlim, col = hcl.colors(12), iasp = 1)

plot_corcolorbar(col = hcl.colors(20, "Blue-Red", rev = TRUE))
}
\arguments{
\item{zlim}{Limits on the color scale}

\item{col}{Vector of colors}

\item{iasp}{Inverse aspect ratio (larger values will result in wider colorbars)}
}
\description{
Plot a colorbar given limits and a set of colors.
}
\section{Functions}{
\itemize{
\item \code{plot_corcolorbar}: Plot a colorbar for correlations
}}

\author{
John K Best
}
