% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_plots.R
\name{plot_field}
\alias{plot_field}
\title{Plot spatial and spatiotemporal fields}
\usage{
plot_field(
  est,
  gen = NULL,
  mesh = generate_mesh(),
  proj_step = 1,
  par_regex = NULL,
  colorbar = FALSE
)
}
\arguments{
\item{est}{Vector or named list of parameter estimates}

\item{gen}{Named list of generative parameter values (not required)}

\item{mesh}{FEM mesh used for spatial/spatiotemporal fields}

\item{proj_step}{Size of projection pixels}

\item{par_regex}{Search string for parameter names (e.g. "omega" will
include both \code{omega_n} and \code{omega_w})}

\item{colorbar}{Plot a colorbar? (NOT IMPLEMENTED)}
}
\description{
Projects and plots fields. Can plot just estimated fields, or include
generative fields for comparison purposes.
}
\author{
John K Best
}
