% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_plots.R
\name{plot_popstate}
\alias{plot_popstate}
\title{Plot array of population states (simulated "truth")}
\usage{
plot_popstate(popstate, col = hcl.colors(12), colorbar = TRUE, plot_cols = 5)
}
\arguments{
\item{popstate}{3-d array with years in the third dimension}

\item{colorbar}{Normalize color scale across years and include a colorbar?}

\item{plot_cols}{How many years in each row?}
}
\value{

}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
John K Best
}
