% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatq_result.R
\name{posdefhess}
\alias{posdefhess}
\title{Is the Hessian positive definite?}
\usage{
posdefhess(x)
}
\arguments{
\item{x}{\code{spatq_result} or \code{spatq_simstudyspec}}
}
\value{
Logical
}
\description{
Is the Hessian positive definite?
}
\author{
John K Best
}
