% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare_adfun}
\alias{prepare_adfun}
\title{Make a TMB ADFun}
\usage{
prepare_adfun(data, parameters, map, random, ..., silent = TRUE)
}
\arguments{
\item{data}{Data list, as from \code{prepare_data}}

\item{parameters}{Parameter list, as from \code{prepare_pars}}

\item{map}{A \code{map} list, as from \code{prepare_map}}

\item{random}{A character vector of parameters to be marginalized, as from
\code{prepare_random}}

\item{...}{Additional arguments to pass to \code{MakeADFun}}

\item{silent}{Output TMB progress?}
}
\value{
A TMB ADFun suitable for optimization
}
\description{
Verify data, parameters, and map, then contruct the ADFun.
}
\author{
John Best
}
