% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare_map}
\alias{prepare_map}
\title{Prepare a \code{map}}
\usage{
prepare_map(pars, map_pars)
}
\arguments{
\item{pars}{Parameter list, as from \code{prepare_pars}}

\item{map_pars}{Character vector indicating which parameters to \code{map}}
}
\value{
A \code{map} list suitable for passing to \code{MakeADFun}
}
\description{
Prepare a \code{map} argument for \code{MakeADFun}. Accepts parameter names
and takes care of dependencies such as spatial or random effect
hyperparameters being map'd when their effects are.
}
\author{
John Best
}
