% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_plots.R
\name{proc_colormap}
\alias{proc_colormap}
\title{Colormaps for numbers density and weight per group fields}
\usage{
proc_colormap(parname, n = 12)
}
\arguments{
\item{parname}{String like "omega_n" or "epsilon_w"}

\item{n}{Number of colors in the map}
}
\value{
Vector of colors, as from \code{\link[grDevices]{hcl.colors}}
}
\description{
Default colormaps for numbers density (viridis) and weight per group
(inferno). These two are chosen based on the regexes \code{"_n$"} and
\code{"_w$"} respectively. Otherwise uses plasma.
}
\author{
John K Best
}
