% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_plots.R
\name{project_field}
\alias{project_field}
\title{Project random effects, reshape to array}
\usage{
project_field(parvec, proj, dim)
}
\arguments{
\item{parvec}{Parameter vector (coerced to vector)}

\item{proj}{Projection matrix}

\item{dim}{Output array dimensions (typically \code{c(s1, s2, nyears)})}
}
\value{
Array of dims \code{dim}
}
\description{
Useful for projecting multiple-year random effects to grids for plotting
with \code{image}.
}
\author{
John K Best
}
