% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_paths.R
\name{rdata_path}
\alias{rdata_path}
\title{Construct the relative path to the Rdata fit file}
\usage{
rdata_path(spec)
}
\arguments{
\item{spec}{A \code{\link{spatq_simstudyspec}} object or list with its
elements}
}
\value{
The relative path to the Rdata file
}
\description{
Construct the relative path to the Rdata fit file
}
\author{
John K Best
}
