% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_pars.R
\name{re_par_idx}
\alias{re_par_idx}
\title{General random effects parameter index}
\usage{
re_par_idx(par_name)
}
\arguments{
\item{par_name}{Name of the spatial/spatiotemporal parameter (e.g.
\code{gamm_n}, \code{eta_w} &c.)}
}
\value{
The index of the effect in the parameter vectors
}
\description{
The variance of general random effects are stored in a length-4 vector. This
function takes a name and returns the corresponding index.
}
\author{
John Best
}
