% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_fits.R
\name{read_all_indices}
\alias{read_all_indices}
\title{Load index CSVs for a study}
\usage{
read_all_indices(study, repls, opmods, estmods, root_dir = ".")
}
\arguments{
\item{study}{Simulation study name}

\item{repls}{Replicate range}

\item{opmods}{Operating model range}

\item{estmods}{Estimation model names}

\item{root_dir}{Root directory}
}
\value{
\code{\link[tibble]{tibble}} with indices for all fitted models
}
\description{
Read all fitted index CSVs
}
\author{
John K Best
}
