% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{read_catch}
\alias{read_catch}
\title{Read simulated data}
\usage{
read_catch(repl, sc, root_dir = ".")
}
\arguments{
\item{repl}{Replicate number}

\item{sc}{Scenario; one of "pref", "spat", or "combo"}

\item{root_dir}{Directory to work in; defaults to current working directory}
}
\value{
A \code{tibble} with catch observations
}
\description{
Read in a replicate of data simulated by FisherySim.jl
}
\details{
Requires that catch files are e.g. \code{repl_01/catch_simple.csv}.
}
\author{
John Best
}
