% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_catchability.R
\name{read_catchability}
\alias{read_catchability}
\title{Read catchability}
\usage{
read_catchability(study, repl, opmod, root_dir = ".", base_q = 0.01)
}
\arguments{
\item{study}{Study name}

\item{repl}{Replicate number}

\item{opmod}{Operating model number}

\item{root_dir}{Root directory}

\item{base_q}{Base catchability}
}
\value{
List with \code{surv} and \code{comm} matrices
}
\description{
Read prep file and calculate realized catchability. Uses rocky habitat
covariate for \code{habq} and \code{bycatch} studies and
\code{log_catchability_devs} for other studies.
}
\author{
John K Best
}
