% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_fits.R
\name{read_estpop}
\alias{read_estpop}
\title{Read and reshape population estimates}
\usage{
read_estpop(x, domdim)
}
\arguments{
\item{domdim}{Domain dimension}

\item{spec}{A \code{\link{spatq_simstudyspec}} or \code{\link{spatq_result}}
object}
}
\value{
An array with first two dimensions \code{domdim}
}
\description{
Load estimated population map from a fit and reshape it so that the first
two dimensions represent space and the third represents time.
}
\author{
John K Best
}
