% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_fits.R
\name{read_index}
\alias{read_index}
\title{Read an index from a fitted model}
\usage{
read_index(x, filetype, estmods)
}
\arguments{
\item{x}{Path to index or \code{\link{spatq_simstudyspec}}}

\item{filetype}{File type, either "feather" or "csv"}

\item{estmods}{Names of estimation models}
}
\value{
A data frame with fitted index values
}
\description{
Read an index from a fitted model
}
\author{
John K Best
}
