% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_fits.R
\name{read_index_csv}
\alias{read_index_csv}
\title{Read indices of abundance}
\usage{
read_index_csv(
  filestudy,
  repl = NULL,
  opmod = NULL,
  estmod = NULL,
  root_dir = ".",
  estmods = c("survey", "spatial_ab", "spatial_q")
)
}
\arguments{
\item{filestudy}{Either the file path or the study name}

\item{repl}{Replicate number}

\item{opmod}{Operating model number}

\item{estmod}{Estimation model name}

\item{root_dir}{Root directory}

\item{estmods}{Names of estimation models to code column as a factor}
}
\value{
\code{\link[tibble]{tibble}} with true and estimated indices of
abundance
}
\description{
Read a CSV containing the true and estimated indices of abundance.
}
\details{
Includes true, estimated, and debiased estimates as well as error estimates.
}
\author{
John K Best
}
