% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_fits.R
\name{read_rdata}
\alias{read_rdata}
\title{Read saved Rdata from a fitted model}
\usage{
read_rdata(x)
}
\arguments{
\item{x}{Path to index, \code{\link{spatq_simstudyspec}}, or list}
}
\value{
A data frame with fitted index values
}
\description{
Read saved Rdata from a fitted model
}
\author{
John K Best
}
