% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_paths.R
\name{repl_dir}
\alias{repl_dir}
\title{Directory name for each replicate}
\usage{
repl_dir(repl)
}
\arguments{
\item{repl}{Replicate number}
}
\value{
A string with the number left-padded to two characters and
concatenated with \code{}"repl_"}
}
\description{
Convenience function for writing replicate directory names
}
\author{
John K Best
}
