% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_paths.R
\name{create_res_dir}
\alias{create_res_dir}
\alias{res_file_names}
\alias{res_file_paths}
\alias{all_res_file_paths}
\title{Paths to each results file}
\usage{
create_res_dir(study, repl = NULL, root_dir = ".")

res_file_names(study, opmod, estmod)

res_file_paths(study, repl, opmod, estmod, root_dir = ".")

all_res_file_paths(study, repls, opmods, estmods, root_dir = ".")
}
\arguments{
\item{study}{Study name}

\item{repl}{Replicate number(s)}

\item{root_dir}{Root directory}

\item{opmod}{Operating model}

\item{estmod}{Estimation model}
}
\value{
A list with elements \code{rdata} and \code{indexcsv} with relative
paths starting from \code{root_dir}
}
\description{
Get paths for the Rdata and index CSV files resulting from a model fit. For
\code{all_res_file_paths}, paths will be returned sorted by replicate number
then operating model, then estimation model.
}
\section{Functions}{
\itemize{
\item \code{create_res_dir}: Create a results directory if necessary

\item \code{res_file_names}: File names for each study, operating, and
estimation model combination

\item \code{all_res_file_paths}: Get paths for all results files
}}

\author{
John K Best
}
