% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_sims.R
\name{rescale_index}
\alias{rescale_index}
\title{Rescale indices}
\usage{
rescale_index(b, sd = NULL)
}
\arguments{
\item{b}{Unscaled indices/biomass}

\item{sd}{Unscaled index standard deviation}
}
\value{
A list with elements \code{index} and \code{sd}, where indices are
scaled to have a product 1 (mean log of zero) and standard deviations are
scaled by the same factor.
}
\description{
Rescale indices of abundance so that they have product one (mean-log zero).
This facilitates comparison among models that only estimate relative
abundance and with simulated true abundances
}
\author{
John Best
}
