% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_sims.R
\name{rmse_metric}
\alias{rmse_metric}
\title{Root mean square error}
\usage{
rmse_metric(est_index, true_index)
}
\arguments{
\item{est_index}{Estimated rescaled index}

\item{true_index}{True rescaled index}
}
\value{
Root mean square error of the estimates
}
\description{
Calculate the root mean square error of the estimated indices of abundance
compared to the true, simulated values. Expects that these will be on a
common scale, e.g. by using \code{rescale_indices}.
}
\author{
John Best
}
