% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_paths.R
\name{sim_file_names}
\alias{sim_file_names}
\alias{sim_file_paths}
\title{Paths to simulated data sets}
\usage{
sim_file_names(study, opmod)

sim_file_paths(study, repl, opmod, root_dir = ".")
}
\arguments{
\item{study}{Study}

\item{opmod}{Operating model}

\item{repl}{Replicate number}

\item{root_dir}{Where to start}
}
\value{
A list of file paths with elements \code{catch_csv},
\code{catch_feather}, \code{pop_csv}, \code{pop_feather}, and \code{poph5}
}
\description{
Get the paths to each simulation file. ' File names are
}
\section{Functions}{
\itemize{
\item \code{sim_file_names}: The name of each simulation file
}}

\author{
John K Best
}
