% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatq.R
\docType{package}
\name{spatq}
\alias{spatq}
\alias{spatq-package}
\title{Spatially varying catchability for fishery-dependent observations}
\description{
A package for fitting index standardization models that combine
fishery-independent and fishery-dependent observations, where the
fishery-dependent observations are allowed a spatially varying
catchability.
}
