% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_adfun.R
\name{spatq_obj}
\alias{spatq_obj}
\title{Contstruct a spatq objective function}
\usage{
spatq_obj(setup, runSymbolicAnalysis = TRUE, normalize = TRUE, ...)
}
\arguments{
\item{setup}{a \code{\link{spatq_setup}} object}

\item{runSymbolicAnalysis}{Use Metis reorderings? (Requires special
installation of TMB; see documentation.)}

\item{normalize}{Normalize GMRF likelihoods?}

\item{...}{Additional arguments to pass to \code{\link[TMB]{MakeADFun}}}
}
\value{
a spatq_obj object suitable for passing to \code{\link{fit_spatq}}
}
\description{
Prepare an objective function using \code{\link[TMB]{MakeADFun}}.
}
\author{
John K Best
}
