% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{spatq_simsetup}
\alias{spatq_simsetup}
\title{Construct a spatq_setup from a simulated data set}
\usage{
spatq_simsetup(
  repl,
  sc,
  sub_df = NULL,
  root_dir = ".",
  max_T = NULL,
  index_step = 5,
  spec_estd = specify_estimated()
)
}
\arguments{
\item{repl}{Replicate number}

\item{sc}{Scenario; "pref", "spat", or "combo"}

\item{sub_df}{Data frame indicating subsampling strategy; see
\code{\link{subsample_catch}}}

\item{root_dir}{Directory to load data from}

\item{max_T}{Last year of data to include}

\item{index_step}{Step for the index grid}

\item{spec_estd}{List of logicals indicating which parameters are to be
estimated, as output \code{\link{specify_estimated}}}
}
\value{
A list with elements \code{data}, \code{parameters}, \code{map}, and
\code{random}, ready to be used as the arguments to
\code{\link{prepare_adfun}}.
}
\description{
Convenience constructor when data follows simulation study directory
structure. See also \code{\link{new_spatq_setup}}.
}
\author{
John K Best
}
