% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyspec.R
\name{spatq_simstudyspec}
\alias{spatq_simstudyspec}
\alias{new_spatq_simstudyspec}
\title{Simulation study specification}
\usage{
spatq_simstudyspec(speclist, feather = TRUE)

new_spatq_simstudyspec(study, repl, opmod, estmod, sub_df, estd)
}
\arguments{
\item{speclist}{List with \code{study}, \code{repl}, \code{opmod},
\code{estmod}, \code{sub_df}, \code{estd}}

\item{study}{Simulation study name}

\item{repl}{Replicate number}

\item{opmod}{Operating model number}

\item{estmod}{Estimation model name}

\item{sub_df}{Data subsetting data frame}

\item{estd}{Parameter estimation specification}
}
\value{
The same list, but with class \code{spatq_simstudyspec}. No checks are
performed.
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\section{Functions}{
\itemize{
\item \code{new_spatq_simstudyspec}: New simstudyspec
}}

\author{
John K Best
}
