% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_estimated.R
\name{specify_estimated}
\alias{specify_estimated}
\title{Specify which parameters to estimate}
\usage{
specify_estimated(
  beta = TRUE,
  gamma = FALSE,
  omega = FALSE,
  epsilon1 = FALSE,
  lambda = TRUE,
  eta = FALSE,
  phi = FALSE,
  psi1 = FALSE
)
}
\arguments{
\item{beta}{Abundance fixed effects}

\item{gamma}{Abundance iid random effects}

\item{omega}{Abundance spatial effects}

\item{epsilon1}{Abundance spatiotemporal effects (sum-to-zero constrained)}

\item{lambda}{Catchability fixed effects}

\item{eta}{Catchability iid random effects}

\item{phi}{Catchability spatial effects}

\item{psi1}{Catchability spatiotemporal effects (sum-to-zero constrained)}
}
\value{
A list with elements for each random effect process and parameter in
the \code{spatq} model, each with a logical indicating whether it is
(\code{TRUE}) or is not (\code{FALSE}) estimated.
}
\description{
Indicate which parameters should be estimated; conversely, which should be
mapped to their initial values. Each argument can be logical, indicating
whether both the numbers density, weight per group, and their
hyperparameters should be estimated. A list specifying whether the numbers
density, weight per group, and hyperparameters (potentially nested inside
each specific process).
}
\author{
John Best
}
