% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_paths.R
\name{study_dir}
\alias{study_dir}
\alias{study_file_base}
\title{Directory where a study's simulations live}
\usage{
study_dir(study, root_dir = ".")

study_file_base(study)
}
\arguments{
\item{study}{Study name}

\item{root_dir}{Where to start}
}
\value{
Directory path or file base name for study
}
\description{
Each set of simulations is stored in a separate directory; this function
serves to look up the name of that directory based on the name of that
directory.
}
\section{Functions}{
\itemize{
\item \code{study_file_base}: The file base for each study
}}

\author{
John K Best
}
