% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{subsample_catch}
\alias{subsample_catch}
\title{Subsample catch by year}
\usage{
subsample_catch(catch_df, n_df = NULL)
}
\arguments{
\item{catch_df}{Catch data frame to subset}

\item{n_df}{Data frame specifiying how to subset}
}
\value{
A data frame with the same columns, subset as specified
}
\description{
Randomly choose a subset of observations for some or all vessels etc.
}
\details{
Takes subsets of observations by year and other characteristics. Subsetting
is specified as a data frame. Any unspecified levels are not subset.
}
\examples{
\dontrun{
  ## To subset the observations of a single vessel, only specify a number of
  ## observations for that vessel
  subsample_catch(catch_df,
                  tibble(vessel_idx = 2,
                         n = 1000))
  ## To subset the observations for two vessels, specify a number of
  ## observations for each
  subsample_catch(catch_df,
                  tibble(vessel_idx = c(1, 2),
                         n = c(200, 1000)))
}
}
\author{
John Best
}
