% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_spatq.R
\name{terminate_opt}
\alias{terminate_opt}
\title{Test whether optimization should be stopped}
\usage{
terminate_opt(fit, optcontrol)
}
\arguments{
\item{fit}{\code{spatq_fit} object from \code{\link{attach_optdiags}}}

\item{optcontrol}{\code{\link{spatq_optcontrol}} object}
}
\value{
boolean indicating termination (TRUE) or not (FALSE)
}
\description{
Sequentially tests whether maximum number of optimizations has been reached,
if the maximum gradient component is less than \code{optcontrol$grtol},
relative parameter value changes between optimizations is less than
\code{optcontrol$dparrtol}, or relative change in the objective function
value is less than \code{optcontrol$dobjrtol}.
}
\author{
John K Best
}
