% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_sims.R
\name{true_percentile}
\alias{true_percentile}
\title{Percentile of true value in estimated distribution}
\usage{
true_percentile(true_index, est_index, est_sd)
}
\arguments{
\item{true_index}{True rescaled index}

\item{est_index}{Estimated index of abundance}

\item{est_sd}{Estimated standard deviation of index}
}
\value{
The percentile of the true value
}
\description{
In order to assess calibration of the estimation models, find the percentile
of the true index under the estimated distribution, assumed to be normal
with estimated mean and standard deviation.
}
\author{
John Best
}
