% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{update_onepar}
\alias{update_onepar}
\title{Update a single parameter vector}
\usage{
update_onepar(currvals, newvals = NULL, mapvec = NULL)
}
\arguments{
\item{currvals}{Vector of current parameter values}

\item{newvals}{Values to update}

\item{mapvec}{Map, indexing the new values into their places in the
parameter vector, with NAs for elements that were not estiamted}
}
\value{
The updated parameter vector
}
\description{
Replace parameter values with better ones, if available.
}
\author{
John K Best
}
