% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{update_setup}
\alias{update_setup}
\alias{update_parameters}
\alias{update_map}
\alias{update_random}
\title{Update a model setup}
\usage{
update_setup(setup, obj, newspec)

update_parameters(setup, obj)

update_map(setup, newspec)

update_random(setup)
}
\arguments{
\item{setup}{Existing setup object}

\item{obj}{Objective function after optimizing}

\item{newspec}{Next estimation specification}
}
\value{
Updated \code{\link{spatqsetup}}
}
\description{
Not intended to be used when the underlying data change.
}
\section{Functions}{
\itemize{
\item \code{update_parameters}: Update \code{parameters}

\item \code{update_map}: Update \code{map}

\item \code{update_random}: Update \code{random}
}}

\author{
John K Best
}
