% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_spatq.R
\name{verify_setequal}
\alias{verify_setequal}
\alias{set_string}
\title{Verify two sets are equal}
\usage{
verify_setequal(set1, set2)

set_string(set)
}
\arguments{
\item{set1}{First set}

\item{set2}{Second set}

\item{set}{A set}
}
\value{
\code{TRUE} if all components present, error otherwise
}
\description{
Check that two arguments contain the same elements. If not, provide an
informative error message.
}
\section{Functions}{
\itemize{
\item \code{set_string}: Paste strings together as a set
}}

