% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_spatq.R
\name{verify_spatq}
\alias{verify_spatq}
\title{Check \code{spatq} inputs}
\usage{
verify_spatq(data, parameters, map = list())
}
\arguments{
\item{data}{Data list}

\item{parameters}{Initial parameter value list}

\item{map}{Parameter map list}
}
\value{
\code{TRUE} if all checks pass
}
\description{
Check data, parameter, and map names, dimensions, and \code{map} specification.
Convenient wrapper around \code{verify_spatq_names}, \code{verify_spatq_dims}, and
\code{verify_spatq_map}.
}
