% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_spatq.R
\name{verify_spatq_dims}
\alias{verify_spatq_dims}
\title{Check \code{spatq} component dimensions}
\usage{
verify_spatq_dims(data, parameters, map = list())
}
\arguments{
\item{data}{Data list for \code{MakeADFun}}

\item{parameters}{Initial parameter value list}

\item{map}{Parameter map list}
}
\value{
\code{TRUE} if all dimension checks pass
}
\description{
Check that all components are of correct dimensions and conform.
}
