% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_spatq.R
\name{verify_spatq_map}
\alias{verify_spatq_map}
\title{Check that \code{map} is self-consistent}
\usage{
verify_spatq_map(parameters, map, warn_not_zero = TRUE)
}
\arguments{
\item{parameters}{Initial parameter value list}

\item{map}{Parameter map list}

\item{warn_not_zero}{Raise a warning if spatial or spatiotemporal components
are map'd off but not all equal to zero.}
}
\value{
\code{TRUE} if all checks pass.
}
\description{
Check that spatial or spatiotemporal field parameters are map'd off if the
corresponding field coefficients are map'd off.
}
