% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_spatq.R
\name{verify_spatq_names}
\alias{verify_spatq_names}
\title{Check all \code{spatq} components are present}
\usage{
verify_spatq_names(data, parameters, map = list())
}
\arguments{
\item{data}{Data list for \code{MakeADFun}}

\item{parameters}{Initial parameter value list}

\item{map}{Parameter map list}
}
\value{
\code{TRUE} if all components present
}
\description{
Check that all the required data and parameter names are included. Currently
doesn't check \code{map} names.
}
