% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_sims.R
\name{bias_metric}
\alias{bias_metric}
\title{Estimate bias of estimation models}
\usage{
bias_metric(fit_df)
}
\arguments{
\item{fit_df}{Data frame with estimated indices, true biomass, and replicate
number}
}
\value{
A list with 'alpha', the offsets for each replicate; 'delta', the
bias coefficient; 'epsilon', the residuals; and 'sigma', the standard
deviation of the regression.
}
\description{
Estimate the relationship between estimated indices and true biomass for
simulated data sets. Uses linear regression to estimate their relationship,
where one indicates unbiasedness. Requires a data frame with columns
'est_index', 'true_biomass', and 'repl'.
}
\details{
.. content for \details{} ..
}
\author{
John Best
}
