% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_spatq.R
\name{fit_spatq}
\alias{fit_spatq}
\alias{report_spatq}
\alias{sdreport_spatq}
\alias{hessian_spatq}
\title{Fit a spatq model by maximum likelihood}
\usage{
fit_spatq(obj, fit = NULL, optcontrol = spatq_optcontrol())

report_spatq(obj)

sdreport_spatq(
  obj,
  bias.correct = !is.null(obj$env$random),
  bias.correct.control = list(sd = TRUE),
  getJointPrecision = FALSE,
  ...
)

hessian_spatq(obj, fit)
}
\arguments{
\item{obj}{A spatq \code{ADFun}, as returned by \code{prepare_adfun} or
\code{make_sim_adfun}}

\item{fit}{Previous fit to use as starting values}

\item{optcontrol}{a}

\item{bias.correct}{Use bias correction for \code{\link[TMB]{sdreport}}?}

\item{getJointPrecision}{Return the joint fixed and random effect precision
matrix from \code{\link[TMB]{sdreport}}?}

\item{...}{additional arguments to pass to \code{\link[TMB]{sdreport}}}
}
\value{
An optimization object, report list, or sdreport list
}
\description{
Uses \code{\link[stats]{optim}} to find an optimum. If a previous fit is not
provided, the first fit uses the conjugate gradient method to improve
quickly. After that, the BFGS method is used to refine the optimization.
Termination tolerances and control parameters for \code{optim} can be
provided through \code{\link{spatq_optcontrol}}.
}
\section{Functions}{
\itemize{
\item \code{report_spatq}: Get object report

\item \code{sdreport_spatq}: Get object sdreport

\item \code{hessian_spatq}: Get finite difference Hessian
}}

\author{
John Best
}
