% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{fixpar_corrplot}
\alias{fixpar_corrplot}
\title{Plot the fixed-effect parameter correlation matrix}
\usage{
fixpar_corrplot(sdr)
}
\arguments{
\item{sdr}{An \code{\link[TMB]{sdreport}} from a TMB model fit}
}
\value{
Nothing; produces plot as a side effect
}
\description{
Plot the fixed-effect parameter correlation matrix, with pretty labels and
clear divisions between parameters. High off-diagonal correlations may
indicate a difficult-to-fit model. Uses the "Blue-Red 3" color palette.
}
\author{
John K Best
}
