% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{generate_mesh}
\alias{generate_mesh}
\title{Generate INLA mesh}
\usage{
generate_mesh()
}
\value{
A \code{inla.mesh} object that can be passed to \code{generate_fem}.
}
\description{
Generate the standard mesh used for simulation fits
}
\author{
John Best
}
