% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{generate_projection}
\alias{generate_projection}
\title{Generate a projection matrix}
\usage{
generate_projection(mesh, data_df, vessel_idx = NULL, group = NULL,
  zero = FALSE)
}
\arguments{
\item{mesh}{INLA mesh to project}

\item{data_df}{Data frame with \code{s1} and \code{s2} columns as
coordinates to project to}

\item{vessel_idx}{Vessel index to project to; others receive zero weight and
are dropped}

\item{group}{Year of observation for spatiotemporal effect}
}
\value{
A sparse projection matrix
}
\description{
Generate a projection matrix to locations in \code{data_df}, possible
zeroing out some observations. Also group by year for a spatiotemporal
effect.
}
\author{
John Best
}
