% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_pars.R
\name{init_fixef}
\alias{init_fixef}
\title{Initial estimates of fixed effect parameters}
\usage{
init_fixef(data)
}
\arguments{
\item{data}{A data list, as produced by \code{prepare_data}}
}
\value{
A list with estimates of \code{beta_n}, \code{beta_w},
\code{lambda_n}, and \code{lambda_w}
}
\description{
Estimates numbers density fixed effects through a binomial regression with
the complementary log-log link function. Log numbers density and log
encounter probability provide an offset for a log normal regression on
positive catches, hopefully providing reasonable starting parameter values
for weight per group covariates. Log numbers density is NOT currently
bias-corrected, but should still work well as starting values.
}
\author{
John Best
}
