% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_estimated.R
\name{is_estd_proc}
\alias{is_estd_proc}
\alias{is_estd_hpar}
\title{Check if process is estimated}
\usage{
is_estd_proc(parname, proc_spec, obs_lik)

is_estd_hpar(parname, estd)
}
\arguments{
\item{parname}{Name of the process parameter (e.g. \code{"omega_n"}) as a string}

\item{proc_spec}{Logical or list at the "process" level (e.g. omega, which
contains \code{omega_n} and \code{omega_w})}

\item{obs_lik}{Observation likelihood, current options are \code{0}:}

\item{estd}{List indicating whether each random process is estimated}
}
\value{
Logical indicating whether parameter is to be estimated
}
\description{
Only intended to be called from \code{\link{specify_estimated}};
\code{proc_spec} argument should take any of the arguments of
\code{specify_estimated} while \code{spec} argument takes a list containing
all of those arguments..
}
\section{Functions}{
\itemize{
\item \code{is_estd_hpar}: Check which hyperparameters are estimated
}}

\author{
John Best
}
