% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_adfun.R
\name{make_sim_adfun}
\alias{make_sim_adfun}
\title{Create ADFun from simulated data}
\usage{
make_sim_adfun(
  study,
  repl,
  opmod,
  estmod = NULL,
  sub_df = NULL,
  spec_estd = NULL,
  root_dir = ".",
  max_T = NULL,
  index_step,
  ...
)
}
\arguments{
\item{study}{Simulation study}

\item{repl}{Replicate number}

\item{opmod}{Operating model}

\item{estmod}{Estimation model}

\item{sub_df}{Data frame indicating subsampling strategy; see
\code{subsample_catch}}

\item{spec_estd}{List of logicals indicating which parameters are to be
estimated, as output \code{\link{specify_estimated}}}

\item{root_dir}{Directory to load data from}

\item{max_T}{Last year of data to include}

\item{index_step}{Index grid step size}

\item{...}{Additional options to pass to \code{prepare_adfun}}
}
\value{
A TMB ADFun suitable for optimization
}
\description{
Read in a simulated data set and construct a TMB ADFun for model fitting.
}
\author{
John Best
}
