% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_maps.R
\name{map_correlation}
\alias{map_correlation}
\title{Calculate the correlation between true and estimated populations}
\usage{
map_correlation(true, est, byyear = FALSE)
}
\arguments{
\item{true}{Array of true biomass}

\item{est}{Array of estimated biomass, with matching first and second
dimensions}

\item{byyear}{Calculate correlations by year?}
}
\value{
Correlation, or vector of correlations by year
}
\description{
Calculates the correlation of true and estimated populations on a regular,
matching grid. It is expected that these are arrays with the first two
dimensions representing space and the third represents time. The
\code{byyear} argument allows for a correlation over the entire result, or
separate correlations for each year.
}
\author{
John K Best
}
