% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{pars_tau}
\alias{pars_tau}
\title{Calculate tau from rho and sigma^2}
\usage{
pars_tau(sig2, rho)
}
\arguments{
\item{sig2}{Marginal variance}

\item{rho}{Correlation range parameter}
}
\value{
The value of tau corresponding to the specified correlation range
and marginal variance
}
\description{
The SPDE construction uses parameters kappa and tau, while the usual Matern
parameters are sigma and rho. The tau parameter is \code{sqrt(1 / (sig2 * 4
* pi * kappa^2))}.
}
\author{
John Best
}
