% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{pars_kappa}
\alias{pars_kappa}
\alias{pars_tau}
\alias{pars_rho}
\alias{pars_sig2}
\title{The SPDE construction uses parameters kappa and tau, while the usual Matern
parameters are sigma and rho. The kappa parameter is \code{sqrt(8) / rho}.}
\usage{
pars_kappa(rho)

pars_tau(sig2, rho)

pars_rho(kappa)

pars_sig2(tau, kappa)
}
\arguments{
\item{rho}{Correlation range parameter}

\item{sig2}{Marginal variance}

\item{kappa}{Correlation decay parameter}

\item{tau}{Ratio of precision and correlation range}
}
\value{
Parameter value
}
\description{
The SPDE construction uses parameters kappa and tau, while the usual Matern
parameters are sigma and rho. The tau parameter is \code{sqrt(1 / (sig2 * 4
* pi * kappa^2))}. All assume that the Matern smoothness parameter is 1.
}
\section{Functions}{
\itemize{
\item \code{pars_kappa}: Calculate kappa from rho

\item \code{pars_rho}: Calculate the correlation range

\item \code{pars_sig2}: Calculate the marginal variance
}}

\author{
John Best
}
