% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare the \code{data} argument for the \code{spatq} model}
\usage{
prepare_data(
  catch_df,
  index_df,
  mesh,
  fem,
  obs_lik = 0L,
  X_contr = contr.helmert
)
}
\arguments{
\item{catch_df}{A data frame with catch observations, as read in by
\code{\link{read_catch}}}

\item{index_df}{A data frame with coordinates for index calculation, as
generated by \code{\link{create_index_df}}}

\item{mesh}{A FEM mesh for the domain}

\item{fem}{The sparse FEM matrices from the \code{mesh} as generated by
\code{generate_fem}}

\item{obs_lik}{Observation likelihood. Current options are \code{0}:
Poisson-link log-normal and \code{1}: Tweedie}

\item{X_contr}{Contrasts to use for fixed effect design matrix}
}
\value{
A \code{list} with the data values for a \code{spatq} model
}
\description{
Create a list of all data elements required to fit a \code{spatq} model with
names etc.
}
\author{
John Best
}
