% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare_map}
\alias{prepare_map}
\title{Prepare a \code{map}}
\usage{
prepare_map(pars, spec)
}
\arguments{
\item{pars}{Parameter list, as from \code{prepare_pars}}

\item{spec}{List of logicals indicating which parameters are to be
estimated, as output by \code{\link{specify_esitmated}}}
}
\value{
A \code{map} list suitable for passing to \code{MakeADFun}
}
\description{
Prepare a \code{map} argument for \code{MakeADFun}.
}
\author{
John Best
}
