% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare_pars}
\alias{prepare_pars}
\title{Prepare parameter list}
\usage{
prepare_pars(data, mesh, init_fixef = TRUE)
}
\arguments{
\item{data}{Data list, as from \code{prepare_data}}

\item{mesh}{SPDE mesh}
}
\value{
List with scalar, vector, or matrices of zeros as starting values
for optimization
}
\description{
Prepare a list of parameters of appropriate dimension. All starting values
are zeros except kappa and tau, which use a correlation range of 50 (half
the domain) and marginal variance of 1, tranformed to the kappa and tau
parameterization.
}
\author{
John Best
}
