% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare_random}
\alias{prepare_random}
\title{Prepare \code{random}}
\usage{
prepare_random(map)
}
\arguments{
\item{map}{A \code{map} list, as from \code{prepare_map}}
}
\value{
A character vector indicating which parameters should be integrated
out using the Laplace approximation.
}
\description{
Prepare a character vector indicating which parameters should be
marginalized out using the Laplace approximation. Does not include
parameters that are \code{map}d.
}
\author{
John Best
}
