% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare_random}
\alias{prepare_random}
\title{Prepare \code{random}}
\usage{
prepare_random(map)
}
\arguments{
\item{map}{A \code{map} list, as from \code{prepare_map}}
}
\value{
A character vector indicating which parameters should be integrated
out using the Laplace approximation.
}
\description{
Prepare a character vector indicating which parameters should be
marginalized out using the Laplace approximation. Does not include
parameters that are \code{map}'d.
}
\details{
Parameters in this model that are typically marginalized are:
\begin{itemize}
\item \code{gamma_n}, \code{gamma_w}
\item \code{omega_n}, \code{omega_w}
\item \code{epsilon_n}, \code{epsilon_w}
\item \code{eta_n}, \code{eta_w}
\item \code{phi_n}, \code{phi_w}
\item \code{psi_n}, \code{psi_w}
\end{itemize}
}
\author{
John Best
}
