% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{read_catch}
\alias{read_catch}
\title{Read simulated data}
\usage{
read_catch(study, repl, opmod, root_dir = ".")
}
\arguments{
\item{study}{Simulation study name}

\item{repl}{Replicate number}

\item{opmod}{Operating model}

\item{root_dir}{Directory to work in; defaults to current working directory}
}
\value{
A \code{tibble} with catch observations
}
\description{
Read in a replicate of data simulated by FisherySim.jl
}
\author{
John Best
}
