% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{read_popstate}
\alias{read_popstate}
\title{Read true population state for each year}
\usage{
read_popstate(
  study,
  repl,
  opmod,
  root_dir = ".",
  filetype = c("feather", "csv", "h5")
)
}
\arguments{
\item{study}{Study name, e.g. "qdevscaling"}

\item{repl}{Replicate number}

\item{opmod}{Operating model}

\item{root_dir}{Directory with e.g. \code{repl_01} subdirectory that
contains \code{popstate_*.csv}}

\item{filetype}{Which popstate file to read? Can be \code{"feather"},
\code{"csv"}, or \code{"h5"}}
}
\value{
A \code{tibble} with population and year, starting from 1
}
\description{
Read true population total from file
}
\author{
John Best
}
