% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_sims.R
\name{rescale_index}
\alias{rescale_index}
\title{Rescale indices}
\usage{
rescale_index(b)
}
\arguments{
\item{b}{Unscaled indices/biomass}
}
\value{
A vector of indices scaled to have a product 1 (mean log of zero)
}
\description{
Rescale indices of abundance so that they have product one (mean-log zero).
This facilitates comparison among models that only estimate relative
abundance and with simulated true abundances
}
\author{
John Best
}
