% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_fits.R
\name{save_index}
\alias{save_index}
\title{Save true and estimated indices to a CSV}
\usage{
save_index(spec, sdr, max_T = 15)
}
\arguments{
\item{spec}{A \code{spatq_spec}}

\item{sdr}{An SD report from \code{\link{sdreport_spatq}}}

\item{max_T}{End year}
}
\value{
The CSV file name
}
\description{
Save the index results, scaled appropriately. Also includes error estimates.
}
\author{
John K Best
}
