% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_paths.R
\name{sim_file_names}
\alias{sim_file_names}
\alias{sim_file_paths}
\title{Paths to simulated data sets}
\usage{
sim_file_names(study, scenario)

sim_file_paths(study, repl, scenario, root_dir = ".")
}
\arguments{
\item{study}{Study; currently only "qdevscaling"}

\item{scenario}{Scenario identifier; "qdevscaling" uses numeric}

\item{repl}{Replicate number}

\item{root_dir}{Where to start}
}
\value{
A list of file paths with elements \code{catch}, \code{popcsv}, and
\code{poph5}
}
\description{
Get the paths to each simulation file.
File names are ${study_dir}/repl_${repl}/${study_scenario}\emph{catch.csv File
names are repl}$repl/catch_$repl_$sc.csv, with $repl padded to two digits.
}
\section{Functions}{
\itemize{
\item \code{sim_file_names}: The name of each simulation file
}}

\author{
John K Best
}
