% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{spatq_simsetup}
\alias{spatq_simsetup}
\title{Construct a spatq_setup from a simulated data set}
\usage{
spatq_simsetup(
  study,
  repl,
  opmod,
  estmod = NULL,
  sub_df = NULL,
  spec_estd = NULL,
  root_dir = ".",
  max_T = NULL,
  index_step = 5,
  init_fixef = TRUE,
  ...
)
}
\arguments{
\item{study}{Simulation study name}

\item{repl}{Replicate number}

\item{opmod}{Operating model}

\item{estmod}{Estimation model}

\item{sub_df}{Data frame indicating subsampling strategy; see
\code{\link{subsample_catch}}}

\item{spec_estd}{List of logicals indicating which parameters are to be
estimated, as output \code{\link{specify_estimated}}}

\item{root_dir}{Directory to load data from}

\item{max_T}{Last year of data to include}

\item{index_step}{Step for the index grid}

\item{init_fixef}{Use parameter estimates from model with only fixed effects
as initial values?}

\item{...}{Additional arguments to pass to \code{\link{init_fixef}}, which
are then passed to \code{\link{spatq_fit}} for the fixed effect model fit}
}
\value{
A list with elements \code{data}, \code{parameters}, \code{map}, and
\code{random}, ready to be used as the arguments to
\code{\link{prepare_adfun}} or passed to \code{\link{spatq_obj}}
}
\description{
Convenience constructor when data follows simulation study directory
structure. See also \code{\link{new_spatq_setup}}.
}
\author{
John K Best
}
