% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_estimated.R
\name{specify_estimated}
\alias{specify_estimated}
\title{Specify which parameters to estimate}
\usage{
specify_estimated(
  beta = TRUE,
  gamma = FALSE,
  omega = FALSE,
  epsilon = FALSE,
  lambda = TRUE,
  eta = FALSE,
  phi = FALSE,
  psi = FALSE,
  kappa_map = NULL,
  aniso = FALSE,
  obs_lik = 0L
)
}
\arguments{
\item{beta}{Abundance fixed effects}

\item{gamma}{Abundance iid random effects}

\item{omega}{Abundance spatial effects}

\item{epsilon}{Abundance spatiotemporal effects}

\item{lambda}{Catchability fixed effects}

\item{eta}{Catchability iid random effects}

\item{phi}{Catchability spatial effects}

\item{psi}{Catchability spatiotemporal effects}

\item{kappa_map}{Factor vector of length 8 specifying the map for kappa
parameters}

\item{obs_lik}{Observation likelihood. Current options are \code{0}:
Poisson-link log-normal and \code{1}: Tweedie}
}
\value{
A list with elements for each random effect process and parameter in
the \code{spatq} model, each with a logical indicating whether it is
(\code{TRUE}) or is not (\code{FALSE}) estimated.
}
\description{
Indicate which parameters should be estimated; conversely, which should be
mapped to their initial values. Each argument can be logical, indicating
whether both the numbers density, weight per group, and their
hyperparameters should be estimated. A list specifying whether the numbers
density, weight per group, and hyperparameters (potentially nested inside
each specific process).
}
\author{
John Best
}
