% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{update_setup}
\alias{update_setup}
\alias{update_parameters}
\alias{update_map}
\alias{update_random}
\title{Update a model setup}
\usage{
update_setup(setup, newparobj, newspec)

update_parameters(setup, newparobj, checkdims = TRUE)

update_map(setup, newspec)

update_random(setup)
}
\arguments{
\item{setup}{Existing setup object}

\item{newparobj}{Object containing updated parameter values, to be extracted
using \code{\link{get_newpars}}}

\item{newspec}{Next estimation specification}
}
\value{
Updated \code{\link{spatqsetup}}
}
\description{
Not intended to be used when the underlying data change.
}
\section{Functions}{
\itemize{
\item \code{update_parameters}: Update \code{parameters}

\item \code{update_map}: Update \code{map}

\item \code{update_random}: Update \code{random}
}}

\author{
John K Best
}
